Move RPC to frontend go files
